requirements.txt
pandas
matplotlib
datetime
datetime
argparse
pathlib

copy 

gspread
google-auth

python-dotenv
